% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ANMC0_2.R
\name{ANMC_Gauss}
\alias{ANMC_Gauss}
\title{ANMC estimate for the remainder}
\usage{
ANMC_Gauss(compBdg, problem, delta = 0.4, type = "M", typeReturn = 0,
  verb = 0)
}
\arguments{
\item{compBdg}{total computational budget in seconds.}

\item{problem}{list defining the problem with mandatory fields \itemize{
\item muEq = mean vector of \eqn{X^{q}};
\item sigmaEq = covariance matrix of \eqn{X^q};
\item Thresh = threshold;
\item muEmq = mean vector of \eqn{X^{-q}};
\item wwCondQ = ``weights'' for \eqn{X^{-q} | X^q} [the vector \eqn{\Sigma^{-q,q}(\Sigma^q)^{-1}}];
\item sigmaCondQChol = Cholesky factorization of the conditional covariance matrix \eqn{\Sigma^{-q | q}};
}}

\item{delta}{total proportion of budget assigned to initial estimate (default 0.4), the actual proportion used might be smaller.}

\item{type}{type of excursion: "m", for minimum below threshold or "M", for maximum above threshold.}

\item{typeReturn}{integer chosen between \itemize{
         \item 0 a number with only the probability estimation;
         \item 1 light return: a list with the probability estimator, the variance of the estimator, the vectors of conditional quantities used to obtain m^* and the system dependent parameters;
         \item 2 heavy return: the same list as light return with also the computational times and additional intermediate parameters.
}}

\item{verb}{level of verbosity (0,1 for this function), also sets the verbosity of trmvrnorm_rej_cpp (to verb-1).}
}
\value{
A list containing the estimated probability of excursion, see typeReturn for details.
}
\description{
Asymmetric nested Monte Carlo estimation of \eqn{P(max X^{-q} > Thresh | max X^{q} \le Thresh)} where X is a normal vector.
}
\references{
Azzimonti, D. and Ginsbourger, D. (2016). Estimating orthant probabilities of high dimensional Gaussian vectors with an application to set estimation. Preprint at \href{https://hal.archives-ouvertes.fr/hal-01289126}{hal-01289126}

Dickmann, F. and Schweizer, N. (2014). Faster comparison of stopping times by nested conditional Monte Carlo. arXiv preprint arXiv:1402.0243.

Genz, A. (1992). Numerical computation of multivariate normal probabilities. Journal of Computational and Graphical Statistics, 1(2):141--149.
}

