% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConsEstv0_6.R
\name{conservativeEstimate}
\alias{conservativeEstimate}
\title{Computationally efficient conservative estimate}
\usage{
conservativeEstimate(alpha = 0.95, pred, design, Thresh, pn = NULL,
  type = ">", verb = 1, lightReturn = T, algo = "GANMC")
}
\arguments{
\item{alpha}{probability of conservative estimate.}

\item{pred}{list containing mean vector (pred$mean) and covariance matrix (pred$cov).}

\item{design}{the discretization design for the field [to be removed?]}

\item{Thresh}{threshold, real number.}

\item{pn}{coverage probability function, vector of the same length as pred$mean (if not specified it is computed).}

\item{type}{type of excursion: ">" for excursion above Thresh or "<" for below.}

\item{verb}{level of verbosity, integer from 1--7.}

\item{lightReturn}{boolean for light return.}

\item{algo}{choice of algorithm for computing probabilities ("GANMC", "GMC").}
}
\value{
A list containing the conservative estimate (\code{set}), the Vorob'ev level (\code{lvs}). If \code{lightReturn=FALSE}, it also returns the actual probability of the set (\code{proba}) and the variance of this estimate (\code{vars}).
}
\description{
Computes conservative estimates with two step GANMC procedure for a Gaussian vector. The probability is approximated with a biased low dimensional estimator and the bias is corrected with a MC estimator.
}
\references{
Azzimonti, D. and Ginsbourger, D. (2016). Estimating orthant probabilities of high dimensional Gaussian vectors with an application to set estimation. Preprint at \href{https://hal.archives-ouvertes.fr/hal-01289126}{hal-01289126}

Bolin, D. and Lindgren, F. (2015). Excursion and contour uncertainty regions for latent Gaussian models. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 77(1):85--106.
}

