% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProbaMin.R
\name{ProbaMin}
\alias{ProbaMin}
\title{Probability of exceedance of minimum of Gaussian vector}
\usage{
ProbaMin(cBdg, q, E, Thresh, mu, Sigma, pn = NULL, lightReturn = T,
  method = 2, verb = 0, Algo = "ANMC")
}
\arguments{
\item{cBdg}{computational budget.}

\item{q}{number of active dimensions.}

\item{E}{discretization design for the field (somehow optional?).}

\item{Thresh}{threshold.}

\item{mu}{mean vector.}

\item{Sigma}{covariance matrix.}

\item{pn}{coverage function vector.}

\item{lightReturn}{boolean, if true light return.}

\item{method}{method chosen to select the active dimensions.}

\item{verb}{level of verbosity (0-5), selects verbosity also for ANMC_Gauss (verb-1) and MC_Gauss (verb-1).}

\item{Algo}{choice of algorithm to compute the remainder Rq ("ANMC" or "MC").}
}
\value{
A list containing the probability estimate ($probabilities) and its variance ($variance), if lightReturn=FALSE it also includes the active dimensions ($indQ), the list returned by the MC estimator for Rq ($resRq).
}
\description{
Computes \eqn{P(minZ \ge Thresh)}
Version 1 with choice of algorithm between ANMC_Gauss and MC_Gauss.
The two most expensive parts are computed with the RCpp functions.
}

