% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizeEq.R
\name{selectEq}
\alias{selectEq}
\title{Select active dimensions for small dimensional estimate}
\usage{
selectEq(q, E, Thresh, mu, Sigma, pn = NULL, method = 1)
}
\arguments{
\item{q}{number of active dimensions.}

\item{E}{discretization design for the field.}

\item{Thresh}{threshold.}

\item{mu}{mean vector.}

\item{Sigma}{covariance matrix.}

\item{pn}{coverage function (if NULL it is computed).}

\item{method}{integer chosen between \itemize{
\item 0  selects by taking equally spaced indexes in mu;
\item 1  samples from pn;
\item 2  samples from pn*(1-pn);
\item 3  samples from pn adjusting for the distance (tries to explore all modes);
\item 4  samples from pn*(1-pn) adjusting for the distance (tries to explore all modes);
\item 5  samples with equal probabilities.
}}
}
\value{
A vector of integers denoting the chosen active dimensions of the vector mu.
}
\description{
Select active dimensions for small dimensional estimate
}

