% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbaMin.R
\name{ProbaMin}
\alias{ProbaMin}
\title{Probability of exceedance of minimum of Gaussian vector}
\usage{
ProbaMin(cBdg, q, E, Thresh, mu, Sigma, pn = NULL, lightReturn = T,
  method = 3, verb = 0, Algo = "ANMC")
}
\arguments{
\item{cBdg}{computational budget.}

\item{q}{number of active dimensions.}

\item{E}{discretization design for the field (somehow optional?).}

\item{Thresh}{threshold.}

\item{mu}{mean vector.}

\item{Sigma}{covariance matrix.}

\item{pn}{coverage function vector.}

\item{lightReturn}{boolean, if true light return.}

\item{method}{method chosen to select the active dimensions.}

\item{verb}{level of verbosity (0-5), selects verbosity also for ANMC_Gauss (verb-1) and MC_Gauss (verb-1).}

\item{Algo}{choice of algorithm to compute the remainder Rq ("ANMC" or "MC").}
}
\value{
A list containing
\itemize{
   \item{\code{probability}: }{The probability estimate}
   \item{\code{variance}: }{the variance of the probability estimate}
}
If \code{lightReturn=F} then the list also contains:
\itemize{
   \item{\code{aux_probabilities}: }{ a list with the probability estimates: \code{probability} the actual probability, \code{pq} the biased estimator \eqn{p_q}, \code{Rq} the conditional probability \eqn{R_q}}
   \item{\code{Eq}: }{the points of the design \eqn{E} selected for \eqn{p_q}}
   \item{\code{indQ}: }{the indices of the active dimensions chosen for \eqn{p_q}}
   \item{\code{resRq}: }{The list returned by the MC method used for \eqn{R_q}}
 }
}
\description{
Computes \eqn{P(min X \le Thresh)}
with choice of algorithm between ANMC_Gauss and MC_Gauss.
The two most expensive parts are computed with the RCpp functions.
}
\references{
Azzimonti, D. and Ginsbourger, D. (2016). Estimating orthant probabilities of high dimensional Gaussian vectors with an application to set estimation. Preprint at \href{https://hal.archives-ouvertes.fr/hal-01289126}{hal-01289126}

Chevalier, C. (2013). Fast uncertainty reduction strategies relying on Gaussian process models. PhD thesis, University of Bern.

Dickmann, F. and Schweizer, N. (2014). Faster comparison of stopping times by nested conditional Monte Carlo. arXiv preprint arXiv:1402.0243.

Genz, A. (1992). Numerical computation of multivariate normal probabilities. Journal of Computational and Graphical Statistics, 1(2):141--149.
}

