% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeEq.R
\name{selectEq}
\alias{selectEq}
\title{Select active dimensions for small dimensional estimate}
\usage{
selectEq(q = NULL, E, Thresh, mu, Sigma, pn = NULL, method = 1,
  verb = 0)
}
\arguments{
\item{q}{either the fixed number of active dimensions or the range where the number of active dimensions is chosen with \code{selectQdims}. If \code{NULL} the function \code{selectQdims} is called.}

\item{E}{discretization design for the field.}

\item{Thresh}{threshold.}

\item{mu}{mean vector.}

\item{Sigma}{covariance matrix.}

\item{pn}{coverage function (if NULL it is computed).}

\item{method}{integer chosen between \itemize{
\item 0  selects by taking equally spaced indexes in mu;
\item 1  samples from pn;
\item 2  samples from pn*(1-pn);
\item 3  samples from pn adjusting for the distance (tries to explore all modes);
\item 4  samples from pn*(1-pn) adjusting for the distance (tries to explore all modes);
\item 5  samples with uniform probabilities.
}}

\item{verb}{level of verbosity: 0 returns nothing, 1 returns minimal info}
}
\value{
A vector of integers denoting the chosen active dimensions of the vector mu.
}
\description{
The function \code{selectEq} selects the active dimensions for the computation of \eqn{p_q} with an heuristic method.
}
\references{
Azzimonti, D. and Ginsbourger, D. (2016). Estimating orthant probabilities of high dimensional Gaussian vectors with an application to set estimation. Preprint at \href{https://hal.archives-ouvertes.fr/hal-01289126}{hal-01289126}

Chevalier, C. (2013). Fast uncertainty reduction strategies relying on Gaussian process models. PhD thesis, University of Bern.

Genz, A. (1992). Numerical computation of multivariate normal probabilities. Journal of Computational and Graphical Statistics, 1(2):141--149.
}

