% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trmvrnorm_rej_cpp}
\alias{trmvrnorm_rej_cpp}
\title{Sample from truncated multivariate normal distribution with C++}
\usage{
trmvrnorm_rej_cpp(n, mu, sigma, lower, upper, verb)
}
\arguments{
\item{n}{number of simulations.}

\item{mu}{mean vector.}

\item{sigma}{covariance matrix.}

\item{lower}{vector of lower bounds.}

\item{upper}{vector of upper bounds.}

\item{verb}{level of verbosity: if lower than 3 nothing, 3 minimal, 4 extended.}
}
\value{
A matrix containing of size dxn containing the samples.
}
\description{
Simulates realizations from a truncated multivariate normal with mean mu, covariance matrix sigma in the bounds lower upper.
}

